requirements.txt
fastapi
aiofiles
python-multipart
python-dotenv
uvicorn
openai